Deduplicates tests for sequences
